requirements.txt
django-filter
requests==2.24.0
psycopg2
python-dateutil==2.8.1
pytz==2019.1
Pillow~=8.2
django-cors-headers==3.2.1
whitenoise==5.2.0
drf-yasg==1.20.0
Django==3.1.6
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0